% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_delete_datastore}
\alias{iotanalytics_delete_datastore}
\title{Deletes the specified data store}
\usage{
iotanalytics_delete_datastore(datastoreName)
}
\arguments{
\item{datastoreName}{[required] The name of the data store to delete.}
}
\description{
Deletes the specified data store.
}
\section{Request syntax}{
\preformatted{iotanalytics$delete_datastore(
  datastoreName = "string"
)
}
}

\keyword{internal}
