% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_delete_pipeline}
\alias{iotanalytics_delete_pipeline}
\title{Deletes the specified pipeline}
\usage{
iotanalytics_delete_pipeline(pipelineName)
}
\arguments{
\item{pipelineName}{[required] The name of the pipeline to delete.}
}
\description{
Deletes the specified pipeline.
}
\section{Request syntax}{
\preformatted{iotanalytics$delete_pipeline(
  pipelineName = "string"
)
}
}

\keyword{internal}
